+++ packages/ui/modals/Review.tsx
@@ -1,16 +1,21 @@
-import { Box, type ButtonProps, createPolymorphicComponent, Modal } from '@mantine/core'
+import { Box, type ButtonProps, createPolymorphicComponent, Group, Modal } from '@mantine/core'
 import { useDisclosure } from '@mantine/hooks'
 import { forwardRef } from 'react'
 
+import { Breadcrumb } from '~ui/components/core/Breadcrumb'
 import { UserReviewSubmit } from '~ui/components/core/UserReviewSubmit'
 import { useScreenSize } from '~ui/hooks/useScreenSize'
-
-import { ModalTitle } from './ModalTitle'
 
 const ReviewModalBody = forwardRef<HTMLButtonElement, ReviewModalProps>((props, ref) => {
 	const [opened, handler] = useDisclosure(false)
 	const { isMobile } = useScreenSize()
-	const modalTitle = <ModalTitle breadcrumb={{ option: 'close', onClick: () => handler.close() }} />
+	const modalTitle = (
+		<Group position='apart' align='center' noWrap>
+			<Box maw='70%' style={{ overflow: 'hidden' }}>
+				<Breadcrumb onClick={handler.close} option='close' />
+			</Box>
+		</Group>
+	)
 
 	return (
 		<>
